requirements.txt
click==8.1.7
rich==13.7.0
aiohttp==3.9.1
asyncio==3.4.3
PyGithub==2.1.1
python-gitlab==4.2.0
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
validators==0.22.0
colorama==0.4.6
wcwidth==0.2.13
tqdm==4.66.1
